# storage/app/openapi/index.yaml
openapi: 3.0.0
info:
  title: Memochat API
  description: An API for messenger app
  version: 0.0.1
servers:
  - url: http://localhost
    description: Local server
...
paths:
...
  /api/students/chats:
    get:
      summary: Retrieve student chats
      description: Fetch a list of chats for the authenticated student, including the details of the associated teacher.
      operationId: getStudentChats
      tags:
        - 6. Students
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          description: The page number to fetch.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: A list of chats retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Request successful
                  data:
                    type: object
                    properties:
                      chats:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            student_id:
                              type: integer
                              example: 1
                            teacher_id:
                              type: integer
                              example: 1
                            user:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 1
                                first_name:
                                  type: string
                                  example: Maksym
                                last_name:
                                  type: string
                                  example: Prokopenko
                                email:
                                  type: string
                                  example: teacher@gmail.com
                                avatar_url:
                                  type: string
                                  format: url
                                  example: https://media.istockphoto.com/id/1298261537/ru/векторная/пустой-человек-профиль-глава-заполнитель.jpg
                                role:
                                  type: integer
                                  example: 2
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      total:
                        type: integer
                        example: 1
...
  /api/teachers/chats:
    get:
      summary: Retrieve teacher chats
      description: Fetch a list of chats for the authenticated teacher, including the details of the associated student.
      operationId: getTeacherChats
      tags:
        - 7. Teachers
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          description: The page number to fetch.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: A list of chats retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Request successful
                  data:
                    type: object
                    properties:
                      chats:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            student_id:
                              type: integer
                              example: 1
                            teacher_id:
                              type: integer
                              example: 1
                            user:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  example: 1
                                first_name:
                                  type: string
                                  example: Maksym
                                last_name:
                                  type: string
                                  example: Prokopenko
                                email:
                                  type: string
                                  example: dd@gmail.com
                                avatar_url:
                                  type: string
                                  format: url
                                  example: https://s1ma-diploma-dev.s3.amazonaws.com/users/students/1/avatar_671fc6707c2311.00585783.jpg
                                role:
                                  type: integer
                                  example: 1
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      total:
                        type: integer
                        example: 1
...
  /api/chats/{chat_id}/message:
    post:
      summary: Send a message in a chat
      description: Send a text or media message in a specified chat.
      operationId: sendMessage
      tags:
        - 9. Chats
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the chat where the message is being sent.
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content_type:
                  type: integer
                  enum: [ 1, 2 ]
                  description: |
                    Type of the message content.
                    1 - Text, 2 - Media (image).
                  example: 2
                content:
                  type: string
                  description: |
                    Message content. Should be a string if content_type is text, or a media file (JPEG, PNG, JPG) if content_type is media.
                  oneOf:
                    - type: string
                      maxLength: 4000
                    - type: string
                      format: binary
      responses:
        '200':
          description: Message sent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Request successful
                  data:
                    type: object
                    properties:
                      message:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 17
                          chat_id:
                            type: integer
                            example: 1
                          teacher_id:
                            type: integer
                            nullable: true
                            example: null
                          student_id:
                            type: integer
                            example: 1
                          content:
                            type: string
                            format: url
                            example: "https://s1ma-diploma-dev.s3.amazonaws.com/chats/1/file_67211d075a7cf4.40071370.jpg"
                          content_type:
                            type: integer
                            example: 2
                          created_at:
                            type: integer
                            example: 1730223367
  /api/chats/{chat_id}/messages:
    get:
      summary: Retrieve chat messages
      description: Retrieve a list of messages in a specified chat with pagination.
      operationId: getChatMessages
      tags:
        - 9. Chats
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the chat to retrieve messages from.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            description: The page number for pagination.
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of messages retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Request successful
                  data:
                    type: object
                    properties:
                      chat:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 1
                          student_id:
                            type: integer
                            example: 1
                          teacher_id:
                            type: integer
                            example: 1
                          user:
                            type: object
                            properties:
                              id:
                                type: integer
                                example: 1
                              first_name:
                                type: string
                                example: Maksym
                              last_name:
                                type: string
                                example: Prokopenko
                              email:
                                type: string
                                example: teacher@gmail.com
                              avatar_url:
                                type: string
                                format: url
                                example: "https://media.istockphoto.com/id/1298261537/ru/%D0%B2%D0%B5%D0%BA%D1%82%D0%BE%D1%80%D0%BD%D0%B0%D1%8F/%D0%BF%D1%83%D1%81%D1%82%D0%BE%D0%B9-%D1%87%D0%B5%D0%BB%D0%BE%D0%B2%D0%B5%D0%BA-%D0%BF%D1%80%D0%BE%D1%84%D0%B8%D0%BB%D1%8C-%D0%B3%D0%BB%D0%B0%D0%B2%D0%B0-%D0%B7%D0%BD%D0%B0%D1%87%D0%BE%D0%BA-%D0%B7%D0%B0%D0%BF%D0%BE%D0%BB%D0%BD%D0%B8%D1%82%D0%B5%D0%BB%D1%8C.jpg?s=612x612&w=0&k=20&c=2gTIDAzkqz_kKwQqpo4q4BjimC4M5w98pgISjN3prA8="
                              role:
                                type: integer
                                example: 2
                      messages:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 3
                            chat_id:
                              type: integer
                              example: 1
                            teacher_id:
                              type: integer
                              nullable: true
                              example: null
                            student_id:
                              type: integer
                              example: 1
                            content:
                              type: string
                              format: url
                              example: "https://s1ma-diploma-dev.s3.amazonaws.com/chats/1/file_67211d075a7cf4.40071370.jpg"
                            content_type:
                              type: integer
                              example: 2
                            created_at:
                              type: integer
                              example: 1730137010
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      total:
                        type: integer
                        example: 13
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer